// amplify/functions/generate-camper-pdf/graphql/API.ts
export type GenerateCamperPdfQueryVariables = {
  camperSub: string,
};

export type GenerateCamperPdfQuery = {
  generateCamperPdf?:  {
    __typename: "CamperProfile",
    address?: string | null,
    applicationComplete?: boolean | null,
    applicationFilepath?: string | null,
    arrivedAtCamp?: boolean | null,
    attendanceConfirmations?: number | null,
    birthdate?: string | null,
    camp?:  {
      __typename: "Camp",
      applicationDeadline: string,
      createdAt: string,
      endDate: string,
      id: string,
      medicalFormDeadline?: string | null,
      startDate: string,
      updatedAt: string,
    } | null,
    campId: string,
    city?: string | null,
    createdAt: string,
    dietaryRestrictions?: string | null,
    dietaryRestrictionsNotes?: string | null,
    documents?:  {
      __typename: "ModelCamperDocumentConnection",
      nextToken?: string | null,
    } | null,
    documentsComplete?: boolean | null,
    email: string,
    emergencyContactName?: string | null,
    emergencyContactPhone?: string | null,
    emergencyContactRelationship?: string | null,
    firstName?: string | null,
    gender?: string | null,
    guidanceCounselorEmail?: string | null,
    guidanceCounselorName?: string | null,
    guidanceCounselorPhone?: string | null,
    highSchool?: string | null,
    identityId: string,
    lastName?: string | null,
    middleInitial?: string | null,
    nickname?: string | null,
    owner?: string | null,
    parent1Email?: string | null,
    parent1FirstName?: string | null,
    parent1LastName?: string | null,
    parent1Phone?: string | null,
    parent2Email?: string | null,
    parent2FirstName?: string | null,
    parent2LastName?: string | null,
    parent2Phone?: string | null,
    phone?: string | null,
    profileComplete?: boolean | null,
    recommendation?:  {
      __typename: "ModelRecommendationConnection",
      nextToken?: string | null,
    } | null,
    rotarianReview?:  {
      __typename: "RotarianReview",
      camperUserSub: string,
      createdAt: string,
      owner?: string | null,
      review?: RotarianReviewDecision | null,
      updatedAt: string,
    } | null,
    rotaryClub?:  {
      __typename: "RotaryClub",
      createdAt: string,
      id: string,
      name: string,
      requiresApplication?: boolean | null,
      requiresLetterOfRecommendation?: boolean | null,
      updatedAt: string,
    } | null,
    rotaryClubId?: string | null,
    state?: string | null,
    updatedAt: string,
    userSub: string,
    zipcode?: string | null,
  } | null,
};

export type GetCampQueryVariables = {
  id: string,
};

export type GetCampQuery = {
  getCamp?:  {
    __typename: "Camp",
    applicationDeadline: string,
    camperProfiles?:  {
      __typename: "ModelCamperProfileConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    documentTemplates?:  {
      __typename: "ModelDocumentTemplateConnection",
      nextToken?: string | null,
    } | null,
    endDate: string,
    filterState?:  {
      __typename: "CamperProfileFilterState",
      acceptedCampers?: boolean | null,
      applicationComplete?: boolean | null,
      confirmedCampers?: boolean | null,
      documentsComplete?: boolean | null,
      profileComplete?: boolean | null,
      rejectedCampers?: boolean | null,
    } | null,
    id: string,
    medicalFormDeadline?: string | null,
    startDate: string,
    updatedAt: string,
    viewState?:  {
      __typename: "CamperProfileViewState",
      address?: boolean | null,
      applicationComplete?: boolean | null,
      applicationFilepath?: boolean | null,
      arrivedAtCamp?: boolean | null,
      attendanceConfirmations?: boolean | null,
      birthdate?: boolean | null,
      camp?: boolean | null,
      campId?: boolean | null,
      city?: boolean | null,
      createdAt?: boolean | null,
      dietaryRestrictions?: boolean | null,
      dietaryRestrictionsNotes?: boolean | null,
      documents?: boolean | null,
      documentsComplete?: boolean | null,
      email?: boolean | null,
      emergencyContactName?: boolean | null,
      emergencyContactPhone?: boolean | null,
      emergencyContactRelationship?: boolean | null,
      firstName?: boolean | null,
      gender?: boolean | null,
      guidanceCounselorEmail?: boolean | null,
      guidanceCounselorName?: boolean | null,
      guidanceCounselorPhone?: boolean | null,
      highSchool?: boolean | null,
      identityId?: boolean | null,
      lastName?: boolean | null,
      middleInitial?: boolean | null,
      nickname?: boolean | null,
      parent1Email?: boolean | null,
      parent1FirstName?: boolean | null,
      parent1LastName?: boolean | null,
      parent1Phone?: boolean | null,
      parent2Email?: boolean | null,
      parent2FirstName?: boolean | null,
      parent2LastName?: boolean | null,
      parent2Phone?: boolean | null,
      phone?: boolean | null,
      profileComplete?: boolean | null,
      recommendation?: boolean | null,
      rotarianReview?: boolean | null,
      rotaryClub?: boolean | null,
      rotaryClubId?: boolean | null,
      state?: boolean | null,
      userSub?: boolean | null,
      zipcode?: boolean | null,
    } | null,
  } | null,
};

export type GetCamperDocumentQueryVariables = {
  camperUserSub: string,
  templateId: string,
};

export type GetCamperDocumentQuery = {
  getCamperDocument?:  {
    __typename: "CamperDocument",
    approved?: boolean | null,
    camper?:  {
      __typename: "CamperProfile",
      address?: string | null,
      applicationComplete?: boolean | null,
      applicationFilepath?: string | null,
      arrivedAtCamp?: boolean | null,
      attendanceConfirmations?: number | null,
      birthdate?: string | null,
      campId: string,
      city?: string | null,
      createdAt: string,
      dietaryRestrictions?: string | null,
      dietaryRestrictionsNotes?: string | null,
      documentsComplete?: boolean | null,
      email: string,
      emergencyContactName?: string | null,
      emergencyContactPhone?: string | null,
      emergencyContactRelationship?: string | null,
      firstName?: string | null,
      gender?: string | null,
      guidanceCounselorEmail?: string | null,
      guidanceCounselorName?: string | null,
      guidanceCounselorPhone?: string | null,
      highSchool?: string | null,
      identityId: string,
      lastName?: string | null,
      middleInitial?: string | null,
      nickname?: string | null,
      owner?: string | null,
      parent1Email?: string | null,
      parent1FirstName?: string | null,
      parent1LastName?: string | null,
      parent1Phone?: string | null,
      parent2Email?: string | null,
      parent2FirstName?: string | null,
      parent2LastName?: string | null,
      parent2Phone?: string | null,
      phone?: string | null,
      profileComplete?: boolean | null,
      rotaryClubId?: string | null,
      state?: string | null,
      updatedAt: string,
      userSub: string,
      zipcode?: string | null,
    } | null,
    camperUserSub: string,
    createdAt: string,
    filepath?: string | null,
    owner?: string | null,
    received?: boolean | null,
    template?:  {
      __typename: "DocumentTemplate",
      campId: string,
      createdAt: string,
      filepath?: string | null,
      id: string,
      name?: string | null,
      required?: boolean | null,
      type?: DocumentTemplateType | null,
      updatedAt: string,
    } | null,
    templateId: string,
    updatedAt: string,
  } | null,
};

export type GetCamperProfileQueryVariables = {
  userSub: string,
};

export type GetCamperProfileQuery = {
  getCamperProfile?:  {
    __typename: "CamperProfile",
    address?: string | null,
    applicationComplete?: boolean | null,
    applicationFilepath?: string | null,
    arrivedAtCamp?: boolean | null,
    attendanceConfirmations?: number | null,
    birthdate?: string | null,
    camp?:  {
      __typename: "Camp",
      applicationDeadline: string,
      createdAt: string,
      endDate: string,
      id: string,
      medicalFormDeadline?: string | null,
      startDate: string,
      updatedAt: string,
    } | null,
    campId: string,
    city?: string | null,
    createdAt: string,
    dietaryRestrictions?: string | null,
    dietaryRestrictionsNotes?: string | null,
    documents?:  {
      __typename: "ModelCamperDocumentConnection",
      nextToken?: string | null,
    } | null,
    documentsComplete?: boolean | null,
    email: string,
    emergencyContactName?: string | null,
    emergencyContactPhone?: string | null,
    emergencyContactRelationship?: string | null,
    firstName?: string | null,
    gender?: string | null,
    guidanceCounselorEmail?: string | null,
    guidanceCounselorName?: string | null,
    guidanceCounselorPhone?: string | null,
    highSchool?: string | null,
    identityId: string,
    lastName?: string | null,
    middleInitial?: string | null,
    nickname?: string | null,
    owner?: string | null,
    parent1Email?: string | null,
    parent1FirstName?: string | null,
    parent1LastName?: string | null,
    parent1Phone?: string | null,
    parent2Email?: string | null,
    parent2FirstName?: string | null,
    parent2LastName?: string | null,
    parent2Phone?: string | null,
    phone?: string | null,
    profileComplete?: boolean | null,
    recommendation?:  {
      __typename: "ModelRecommendationConnection",
      nextToken?: string | null,
    } | null,
    rotarianReview?:  {
      __typename: "RotarianReview",
      camperUserSub: string,
      createdAt: string,
      owner?: string | null,
      review?: RotarianReviewDecision | null,
      updatedAt: string,
    } | null,
    rotaryClub?:  {
      __typename: "RotaryClub",
      createdAt: string,
      id: string,
      name: string,
      requiresApplication?: boolean | null,
      requiresLetterOfRecommendation?: boolean | null,
      updatedAt: string,
    } | null,
    rotaryClubId?: string | null,
    state?: string | null,
    updatedAt: string,
    userSub: string,
    zipcode?: string | null,
  } | null,
};

export type GetDocumentTemplateQueryVariables = {
  id: string,
};

export type GetDocumentTemplateQuery = {
  getDocumentTemplate?:  {
    __typename: "DocumentTemplate",
    camp?:  {
      __typename: "Camp",
      applicationDeadline: string,
      createdAt: string,
      endDate: string,
      id: string,
      medicalFormDeadline?: string | null,
      startDate: string,
      updatedAt: string,
    } | null,
    campId: string,
    camperDocuments?:  {
      __typename: "ModelCamperDocumentConnection",
      nextToken?: string | null,
    } | null,
    createdAt: string,
    filepath?: string | null,
    id: string,
    name?: string | null,
    required?: boolean | null,
    type?: DocumentTemplateType | null,
    updatedAt: string,
  } | null,
};

export type GetRecommendationQueryVariables = {
  id: string,
};
// ...
export type GetRotarianProfileQueryVariables = {
  userSub: string,
};

export type GetRotarianProfileQuery = {
  getRotarianProfile?:  {
    __typename: "RotarianProfile",
    approved?: boolean | null,
    createdAt: string,
    email: string,
    firstName?: string | null,
    lastName?: string | null,
    owner?: string | null,
    rotaryClub?:  {
      __typename: "RotaryClub",
      createdAt: string,
      id: string,
      name: string,
      requiresApplication?: boolean | null,
      requiresLetterOfRecommendation?: boolean | null,
      updatedAt: string,
    } | null,
    rotaryClubId?: string | null,
    updatedAt: string,
    userSub: string,
  } | null,
};

export type GetRotarianReviewQueryVariables = {
  camperUserSub: string,
};
// ...
export type GetRotaryClubQueryVariables = {
  id: string,
};
// ...
export type GetUserQueryVariables = {
  username: string,
};

export type GetUserQuery = {
  getUser?: string | null,
};
// ...
export type ListGroupsForUserQueryVariables = {
  username: string,
};